+++ src/internal/data.ts
@@ -3,29 +3,12 @@ import type { Model, SchemaDefinition } from 'mongoose';
 import type { IGlobalOptions, VirtualOptions } from '../types';
 import { Severity } from './constants';
 
-export interface IPreHook {
-  method: string | RegExp;
-  func(error?: Error): void;
-}
-
-export interface IPostHook {
-  method: string | RegExp;
-  func(): void;
-}
-
-export interface IHooks {
-  pre: IPreHook[];
-  post: IPostHook[];
-}
-
 /** Schema Map */
 export const schemas: Map<string, SchemaDefinition> = new Map();
 /** Models Map */
 export const models: Map<string, Model<any>> = new Map();
 /** Virtuals Map */
 export const virtuals: Map<string, Map<string, VirtualOptions>> = new Map();
-/** Hooks Map */
-export const hooks: Map<string, IHooks> = new Map();
 /** Constructors Map */
 export const constructors: Map<string, NewableFunction> = new Map();
 /** Global Options */