+++ src/plugin.ts
@@ -11,13 +11,6 @@ import type { Func, IPluginsArray } from './types';
 export function plugin<T = any>(mongoosePlugin: Func, options?: T) {
   // don't check if options is an object, because any plugin could make it anything
   return (target: any) => {
-    // const name: string = getName(target);
-
-    // /* istanbul ignore else */
-    // if (!plugins.has(name)) {
-    //   plugins.set(name, []);
-    // }
-    // plugins.get(name).push({ mongoosePlugin, options });
     logger.info('Adding plugin "%s" to "%s" with options: "%o"', mongoosePlugin.name, getName(target), options);
     const plugins: IPluginsArray<any>[] = Array.from(Reflect.getMetadata(DecoratorKeys.Plugins, target) ?? []);
     plugins.push({ mongoosePlugin, options });