+++ github-page/changelog.md
@@ -35,6 +35,7 @@ redirect_from:
 - All non-essentail types get exported as `types`
 - `utils`'s `getName` function now gets exported
 - [IC] Move Plugins cache to Reflection
+- [IC] Move Hooks cache to Reflection
 - [IC] All remaining test's models now get exported with an name (not being exported as `model` anymore)
 - Update Dependencies
   - `mongoose` to 5.9.9